Move routes to a different module.
