+++ lib/logging.js
@@ -6,7 +6,7 @@ var winston = require('winston');
  *
  * @param config Configuration object.
  */
-module.exports = function(config) {
+function configure(config) {
     var transports = [];
 
     if(config.persistent) {
@@ -21,5 +21,17 @@ module.exports = function(config) {
     }
 
     return new winston.Logger({transports: transports});
-};
+}
+
+/**
+ * The default logger is a big NOOP.
+ */
+configure.default = function(){};
+configure.default.info = function(){};
+configure.default.debug = function(){};
+configure.default.trace = function(){};
+configure.default.warn = function(){};
+configure.default.error = function(){};
+
+module.exports = configure;
